Same Commented Code From My Previous Test Commit
Became reset during merge. It is draft code for EventsController test
cases copied from the VenuesController test cases. A lot of the
variables are still from venues and will need to be updated. Also, still
need to figure out how to mock out venue objects in order to pass the
object as a String object which the test builder needs.