Added Twitter Instance to EventsControllers
Used the example linked in the week 8 requirements. Still need to verify
if the mapping values need to be adjusted because the the example given
has the root directory as the controller value while our
EventsController is mapped to the "/events" sub-folder. The "/connect"
sub-folder is not under the "/events" path but the "eventsTwitter"
handler method redirects to a page within "/connect" since I directly
copied the linked example provided in the week 8 requirements.